Moved all startup logic to new program.cs
