+++ AuraScheduler.UI/AuraScheduler.UI.csproj
@@ -11,6 +11,7 @@
 		<PlatformTarget>x64</PlatformTarget>
 		<Platforms>x64</Platforms>
 		<ValidateExecutableReferencesMatchSelfContained>false</ValidateExecutableReferencesMatchSelfContained>
+		<StartupObject>AuraScheduler.UI.Program</StartupObject>
 	</PropertyGroup>
 
 	<ItemGroup>
